requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
bcrypt==4.1.2
google-generativeai==0.3.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
agora-token-builder==1.0.0
